requirements.txt
scikit-learn
pandas
joblib
streamlitpush